Time: 1 ms (100.00%), Space: 42.5 MB (37.26%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.java
@@ -14,10 +14,10 @@
  * }
  */
 class Solution {
-    public Map<Integer, Long> m = new HashMap<>();
-    public long ans = 0;
+    public Map<Integer, Integer> m = new HashMap<>();
+    public int ans = 0;
     
-    public void dfs(TreeNode c, int d, long n) {
+    public void dfs(TreeNode c, int d, int n) {
         if (c == null) {
             return;
         }
@@ -33,6 +33,6 @@ class Solution {
     
     public int widthOfBinaryTree(TreeNode root) {
         dfs(root, 1, 1);
-        return (int) ans;
+        return ans;
     }
 }
